lambda/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
Deprecated==1.2.13
docopt==0.6.2
idna==3.4
opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-semantic-conventions==0.37b0
pipreqs==0.4.11
pyaml==21.10.1
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.1
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==1.26.13
wrapt==1.14.1
yarg==0.1.9
